Add login history view and functionality
This commit introduces a new view to display users' login history as well as related changes in the routes and a new livewire component. A `HistoryLogin` file has been created in `app/Livewire/Account` for implementing this view. The layout for this has been added in `resources/views/livewire/account/history-login.blade.php`. A route `account.history.login` has been added in `routes/web.php` to navigate to this page.

Changes also include a new `UserDeviceTable` livewire component and corresponding blade view to display the login history in a tabular format. The header navigation also updated to include a direct link to this new login history page. The necessity of these changes lies in the requirement of providing users the ability to track their sign-in activities for security purposes.