Updated composer.lock and added "log-viewer" dependency
For more effective logging, the ARCANEDEV "Log-Viewer" library was added as a dependency in the composer.json file, and consequently, an update in the composer.lock was needed to reflect this change.

Additionally, a status indication was embedded within the "Authenticator Status" in the account start page, indicating whether this feature is active for the user's account. A new route for "service.otp" was also added alongside its associated Livewire component.

These changes provide better insights on logging and service access for users and offer enhanced OTP service routing within the application.