Add SSO support and UUID to users
Implemented Single Sign On (SSO) support and added unique identifier (UUID) to users. Added a provider attribute to the Livewire login component to handle redirection after login based on the provider. If the provider is 'lab', it redirects to the lab subdomain with logged-in user's UUID as part of the URL. If any other provider, it redirects to the account index route as previously. A new SSO Controller has been added to handle the SSO logic. The User model and migration have been updated to include a unique 'uuid' field. Changes also extend to the middleware to handle API authentication redirections.