Add UserProfil model and update user creation flow
Created a new UserProfil model for storing user profile information. This is reflected in a new migration file which defines the UserProfil table schema. The user creation process has been shifted to a UserObserver object, which triggers when a new user is registered. This change offloads user creation preparation tasks such as service creation and log recording from the registration component and helps improve code separation. The User model has been updated with a one-to-one relationship function for UserProfil and the login component is updated to handle provider setting when mounted. Improvements also include changes to the ProfilController to handle different user querying cases. These changes aim to improve code readability and user-related data management.