Improve User and UserProfil relationships and default values
The changes in the User and UserProfil models allow for better data consistency when retrieving related data. Explicit foreign keys have now been defined to maintain a clear relationship between Users and UserProfil. Additionally, default values have been set in the UserProfil migration to ensure that null values are avoided in the database, improving data integrity. The changes in the LoginController and ProfilController ensure that related data is loaded in one query which will reduce the number of database calls thereby enhancing the performance. In the LoginController, related user data (social, notifications, services, and logs) are now loaded on user sign-in, enabling quick user data access.