Add WikiCategories table and model
Created new migration file for 'wiki_categories' table and its related model.
The table includes columns for the category's 'name', 'icon' and a foreign key 'cercle_id' that references the 'cercles' table. The connection to the related Social Circle model is managed through the 'belongsTo' relation on the WikiCategory model. Similarly, the relationship with 'subcategories' is managed through the 'hasMany' relation. This will facilitate the management and retrieval of Wiki data with respect to different Social Circles and promote clearer categorization.