Add image resize service and avatar update functionality
Added functionality to resize uploaded images and update user avatars. Introduced a new image service and a resize image job with methods to resize and delete the initial file. Updated the account app component with methods to select the avatar form and change the avatar. Included a change avatar form in the account app blade view. Adjusted relevant file paths in the gitignore file and makefile for avatar storage and the deployment process.