Update image cropping size, avatar field, composer.lock and Add Gravatar package
Increased the image cropping size from 100x100 to 512x512 in the Image service to improve clarity. Also added 'avatar' as a new field to the User model to store profile picture data. Updated the composer.lock and composer.json after adding the Gravatar package for user's avatar handling. Inserted the Gravatar usage in Livewire for getting the user's avatar with Gravatar's help.